Add BufferID to memory property to be queried Issue: https://gitlab.devtools.intel.com/one-api/level_zero/issues/105

+++ core/include/xe_memory.inl
@@ -211,10 +211,10 @@ namespace xe
     ///     - **cuPointerGetAttribute**
     /// 
     /// @returns
-    ///     - uint32_t: Value of the queried property
+    ///     - uint64_t: Value of the queried property
     /// 
     /// @throws result_t
-    inline uint32_t 
+    inline uint64_t 
     MemGetProperty(
         mem_allocator_handle_t hMemAllocHandle,         ///< [in] handle of memory allocator for this allocation
         const void* ptr,                                ///< [in] Pointer to query
